Variant: naive support for GT field
Simple check on the PL field whether or not hte sample has PL == 0 in
the queried location. Only works for bi-allelic sites. This is a rough
first implementation to enable other people to start working on their
tools. We will change the underlying impleentation to actually check the
GT field soon.