Variant: make qual float and handle missing values
The missing values in htslib are represented as special bit patterns and
the user is given some macros to test if the result they just got is a
missing value or not. Gamgee will provide a general function
gamgee::is_missing that has overloads appropriate to all value types in
the library. Here we implement the float overload that works for the
qual field specifically.