Variant: simpler template  parameters of select_if
select_if is a wonderful tool, but currently it is forcing the user to
know what is the underlying type so he can use the appropriate template
parameter. This defeats the purpose of creating complicated underlying
types for performance and relying on the user using auto everywhere.

This commit enables select_if to be used as it was intended, without any
template parameters. All template parameters are now fully deduced from
the function parameters.

fixes #153