Travis: add precompiled packages for our tools
* add doxygen 1.8 binary package
* add clang-3.5 + clang modernize
* add gcc 4.9

I also have a boost package compiled with gcc 4.9 but the current clang
package depends on gcc 4.8 (for stdlibc++) and resolving this was a
dependency nightmare that I decided to postpone. Boost-155 is currently
being used from a PPA.

I also added clang-format and clang-modernize #56 but I will only close
that item when we actually use them in the build process. It's available
now.

closes #57