Add option to install dependencies
When building a project that depends on Gamgee, we want to make sure we
also expose any libraries required to build and link with Gamgee. This
adds an option to install those dependencies during the build.

Note that this could conflict with existing packages if we're
installing to a system location, so it is off by default. This is
intended for use only by projects that depend on Gamgee.