Modify htslib.cmake to inherit number of jobs
Calling make using the $(MAKE) built-in variable will inherit the
number of simultaneous jobs to run when the parent project is being
built with make.

If the parent uses a different build system, because of the lack of job
control, building htslib in parallel is likely to be counter-productive
as the parent build system is likely running in parallel already.